requirements.txt
# Setup after virtual environment creation
fastapi[standard]
uvicorn[standard]
SQLAlchemy
alembic
psycopg2
python-jose[cryptography]
passlib[bcrypt]
python-docx
pdfkit
